# docs/swagger.yaml
basePath: /
definitions:
  Entities.ComparisonPair:
    properties:
      comparisonColumnIndex:
        type: integer
      sourceColumnIndex:
        type: integer
    type: object
  recon_requests.GetFileUploadParametersRequest:
    properties:
      comparisionFileName:
        description: comparison file meta data
        type: string
      comparisonFileHash:
        type: string
      comparisonFileRowCount:
        type: integer
      comparisonPairs:
        items:
          $ref: '#/definitions/Entities.ComparisonPair'
        type: array
      sourceFileHash:
        type: string
      sourceFileName:
        type: string
      sourceFileRowCount:
        type: integer
      userId:
        type: string
    required:
    - comparisionFileName
    - comparisonFileHash
    - comparisonPairs
    - sourceFileHash
    - sourceFileName
    - sourceFileRowCount
    - userId
    type: object
  recon_requests.StreamFileChunkForReconRequest:
    properties:
      chunkSequenceNumber:
        type: integer
      fileType:
        type: string
      isEOF:
        type: boolean
      records:
        items:
          type: string
        type: array
      uploadRequestId:
        type: string
    required:
    - chunkSequenceNumber
    - fileType
    - isEOF
    - records
    - uploadRequestId
    type: object
  recon_responses.GetFileUploadParametersResponse:
    properties:
      comparisonFileExpectedBatchSize:
        type: integer
      comparisonFileHash:
        type: string
      comparisonFileLastRowReceived:
        type: integer
      comparisonFileName:
        type: string
      isFirstTimeUploadForCmpFile:
        type: boolean
      isFirstTimeUploadForSourceFile:
        type: boolean
      sourceFileExpectedBatchSize:
        type: integer
      sourceFileHash:
        type: string
      sourceFileLastRowReceived:
        type: integer
      sourceFileName:
        type: string
      uploadRequestId:
        type: string
    type: object
  recon_responses.StreamFileChunkForReconResponse:
    properties:
      messageId:
        type: string
      originalChunkSequenceNumber:
        type: integer
      originalFileType:
        type: string
      status:
        type: string
    type: object
host: us-central1-reconcilercore.cloudfunctions.net
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the core backend Reconciler API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Reconciler Backend Core API
  version: "1.0"
paths:
  /GetFileUploadParameters:
    post:
      consumes:
      - application/json
      description: Receives either Source or Comparison File Chunks and routes them
        appropriately for Reconciliation
      parameters:
      - description: StreamFileChunkForReconRequest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/recon_requests.StreamFileChunkForReconRequest'
      produces:
      - application/json
      responses:
        "200":
          description: StreamFileChunkForReconResponse
          schema:
            $ref: '#/definitions/recon_responses.StreamFileChunkForReconResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: StreamFileChunksForRecon
      tags:
      - StreamFileChunksForRecon API
  /Swagger/index.html:
    get:
      description: used to access the swagger GUI
      produces:
      - application/json
      responses:
        "200":
          description: Json data
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Swagger API
      tags:
      - Swagger APIs
  /Swagger/swagger.json:
    get:
      description: returns json needed by Swagger
      produces:
      - application/json
      responses:
        "200":
          description: Json data
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GetSwaggerJson
      tags:
      - Swagger APIs
swagger: "2.0"